+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## Major changes & new features
 
+- Renamed `strataCheck()` to `bagCheck()` and renamed `strata.R` to `bags.R`
+    - `bagCheck()` was broken down into 2 new minor internal functions (`summarizeBadBags()` and `filterOutOregonPermits()`)
+    - Standardized using the phrase "bag values" over the term "strata"; HIP records contain bag values until they are written out and strata are assigned.
 - Deleted demographic-oriented `outOfStateHunters()` and `youthHunters()` functions because they are not used
 - Almost all data params changed to be less ambiguous (e.g. `x` now `cleaned_data`, `proofed_data`, etc)
 - Functions renamed to camel case except `read_hip()` and `write_hip()`:
@@ -18,7 +21,7 @@
         - New `duplicatePlot()` function added
         - All functions related to duplicates moved to `duplicates.R` (previously separated into `findDuplicates.R` and `fixDuplicates.R`)
     - `read_hip()` was broken down into 15 new minor internal functions (`listFiles()`, `ignorePermits()`, `ignoreHolds()`, `idBlankFiles()`, `dropBlankFiles()`, `checkFileNameDateFormat()`, `checkFileNameStateAbbr()`, `readMessages()`, `missingPIIMessage()`, `missingEmailsMessage()`, `zeroBagsMessage()`, `naBagsMessage()`, `nonDigitBagsMessage()`, `dlStateNAMessage()`, and `dlDateNAMessage()`). More strict requirements must be met for data to be successfully read (e.g. instead of returning a message that file names are incorrectly formatted, this would stop the process). The `zeroBagsMessage()` internal function is a new feature of `read_hip()` that checks for records with all-zero bag values and returns a message to the console if they are detected.
-    - `clean()` was broken down into 10 minor internal functions (1 previously used: `strataFix()` renamed as `permitStrataFix()`; and 9 new functions: `namesToUppercase()`, `nonDigitBagsFilter()`, `naAndZeroBagsFilter()`, `missingPIIFilter()`, `moveSuffixes()`, `fixMiddleInitials()`, `formatZip()`, `zipCheck()`, and `specialOregonHuntYCheck()`)
+    - `clean()` was broken down into 10 minor internal functions (1 previously used: `strataFix()` renamed as `permitBagFix()`; and 9 new functions: `namesToUppercase()`, `nonDigitBagsFilter()`, `naAndZeroBagsFilter()`, `missingPIIFilter()`, `moveSuffixes()`, `fixMiddleInitials()`, `formatZip()`, `zipCheck()`, and `specialOregonHuntYCheck()`)
     - `correct()` 
         - Broken down into 3 minor internal functions (`correctEmail()`, `correctTitle()`, `correctSuffix()`)
         - No longer filters out all-0 bag records (this now happens in `clean()` via `naAndZeroBagsFilter()`)